Create deployment.json
Configuration for redeploying the cluster when changes are made in a git repository on a selected branch. The deploytool only redeploy nodes that have been modified. The deploytool is able to reboot all nodes in the cluster without downtime or data loss by *awaiting* replication before shutting old nodes down. Nodes in the loading state should not register for any dynamic routing within the cluster.